refactor: unify accounts controller
